+++ modules/apps/layout/layout-page-template-item-selector-api/src/main/java/com/liferay/layout/page/template/item/selector/LayoutPageTemplateCollectionTreeNodeItemSelectorCriterion.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.layout.page.template.item.selector.criterion;
+package com.liferay.layout.page.template.item.selector;
 
 import com.liferay.item.selector.BaseItemSelectorCriterion;
